updating data models on the client side
