+++ Sources/KlaviyoCore/InternalAPIModels.swift
@@ -19,212 +19,3 @@ public enum KlaviyoEndpoint: Equatable, Codable {
 extension KlaviyoAPI {
     public static let _appContextInfo = analytics.appContextInfo()
 }
-
-extension PublicProfile.Location: Codable {
-    public init(from decoder: Decoder) throws {
-        let values = try decoder.container(keyedBy: CodingKeys.self)
-        address1 = try values.decode(String.self, forKey: .address1)
-        address2 = try values.decode(String.self, forKey: .address2)
-        city = try values.decode(String.self, forKey: .city)
-        latitude = try values.decode(Double.self, forKey: .latitude)
-        longitude = try values.decode(Double.self, forKey: .longitude)
-        region = try values.decode(String.self, forKey: .region)
-        self.zip = try values.decode(String.self, forKey: .zip)
-        timezone = try values.decode(String.self, forKey: .timezone)
-        country = try values.decode(String.self, forKey: .country)
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(address1, forKey: .address1)
-        try container.encode(address2, forKey: .address2)
-        try container.encode(city, forKey: .city)
-        try container.encode(latitude, forKey: .latitude)
-        try container.encode(longitude, forKey: .longitude)
-        try container.encode(region, forKey: .region)
-        try container.encode(zip, forKey: .zip)
-        try container.encode(timezone, forKey: .timezone)
-        try container.encode(country, forKey: .country)
-    }
-
-    enum CodingKeys: CodingKey {
-        case address1
-        case address2
-        case city
-        case country
-        case latitude
-        case longitude
-        case region
-        case zip
-        case timezone
-    }
-}
-
-public struct PublicEvent: Equatable {
-    public enum EventName: Equatable {
-        case OpenedPush
-        case OpenedAppMetric
-        case ViewedProductMetric
-        case AddedToCartMetric
-        case StartedCheckoutMetric
-        case CustomEvent(String)
-    }
-
-    public struct Metric: Equatable {
-        public let name: EventName
-
-        public init(name: EventName) {
-            self.name = name
-        }
-    }
-
-    struct Identifiers: Equatable {
-        public let email: String?
-        public let phoneNumber: String?
-        public let externalId: String?
-        public init(email: String? = nil,
-                    phoneNumber: String? = nil,
-                    externalId: String? = nil) {
-            self.email = email
-            self.phoneNumber = phoneNumber
-            self.externalId = externalId
-        }
-    }
-
-    public let metric: Metric
-    public var properties: [String: Any] {
-        _properties.value as! [String: Any]
-    }
-
-    private let _properties: AnyCodable
-    public var time: Date
-    public let value: Double?
-    public let uniqueId: String
-    let identifiers: Identifiers?
-
-    init(name: EventName,
-         properties: [String: Any]? = nil,
-         identifiers: Identifiers? = nil,
-         value: Double? = nil,
-         time: Date? = nil,
-         uniqueId: String? = nil) {
-        metric = .init(name: name)
-        _properties = AnyCodable(properties ?? [:])
-        self.time = time ?? analytics.date()
-        self.value = value
-        self.uniqueId = uniqueId ?? analytics.uuid().uuidString
-        self.identifiers = identifiers
-    }
-
-    public init(name: EventName,
-                properties: [String: Any]? = nil,
-                value: Double? = nil,
-                uniqueId: String? = nil) {
-        metric = .init(name: name)
-        _properties = AnyCodable(properties ?? [:])
-        identifiers = nil
-        self.value = value
-        time = analytics.date()
-        self.uniqueId = uniqueId ?? analytics.uuid().uuidString
-    }
-}
-
-public struct PublicProfile: Equatable {
-    public enum ProfileKey: Equatable, Hashable, Codable {
-        case firstName
-        case lastName
-        case address1
-        case address2
-        case title
-        case organization
-        case city
-        case region
-        case country
-        case zip
-        case image
-        case latitude
-        case longitude
-        case custom(customKey: String)
-    }
-
-    public struct Location: Equatable {
-        public var address1: String?
-        public var address2: String?
-        public var city: String?
-        public var country: String?
-        public var latitude: Double?
-        public var longitude: Double?
-        public var region: String?
-        public var zip: String?
-        public var timezone: String?
-        public init(address1: String? = nil,
-                    address2: String? = nil,
-                    city: String? = nil,
-                    country: String? = nil,
-                    latitude: Double? = nil,
-                    longitude: Double? = nil,
-                    region: String? = nil,
-                    zip: String? = nil,
-                    timezone: String? = nil) {
-            self.address1 = address1
-            self.address2 = address2
-            self.city = city
-            self.country = country
-            self.latitude = latitude
-            self.longitude = longitude
-            self.region = region
-            self.zip = zip
-            self.timezone = timezone ?? analytics.timeZone()
-        }
-    }
-
-    public let email: String?
-    public let phoneNumber: String?
-    public let externalId: String?
-    public let firstName: String?
-    public let lastName: String?
-    public let organization: String?
-    public let title: String?
-    public let image: String?
-    public let location: Location?
-    public var properties: [String: Any] {
-        _properties.value as! [String: Any]
-    }
-
-    let _properties: AnyCodable
-
-    public init(email: String? = nil,
-                phoneNumber: String? = nil,
-                externalId: String? = nil,
-                firstName: String? = nil,
-                lastName: String? = nil,
-                organization: String? = nil,
-                title: String? = nil,
-                image: String? = nil,
-                location: Location? = nil,
-                properties: [String: Any]? = nil) {
-        self.email = email
-        self.phoneNumber = phoneNumber
-        self.externalId = externalId
-        self.firstName = firstName
-        self.lastName = lastName
-        self.organization = organization
-        self.title = title
-        self.image = image
-        self.location = location
-        _properties = AnyCodable(properties ?? [:])
-    }
-}
-
-extension PublicEvent.EventName {
-    public var value: String {
-        switch self {
-        case .OpenedPush: return "$opened_push"
-        case .OpenedAppMetric: return "Opened App"
-        case .ViewedProductMetric: return "Viewed Product"
-        case .AddedToCartMetric: return "Added to Cart"
-        case .StartedCheckoutMetric: return "Started Checkout"
-        case let .CustomEvent(value): return "\(value)"
-        }
-    }
-}
